requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
mysql-connector-python>=8.0.33
pymysql>=1.1.0
python-dateutil>=2.8.2
openai>=1.3.0
beautifulsoup4>=4.12.2
playwright>=1.42.0
PyMuPDF>=1.23.7

